requirements.txt
pyTelegramBotAPI
mysql-connector-python
python-dotenv